// src/lib/enums/product.enum.ts
export enum ProductStatus {
    PAUSE = 'PAUSE',
    PROCESS = 'PROCESS',
    DELETE = 'DELETE',
    ACTIVE = "ACTIVE",
}

export enum ProductCollection {
    FILLER = 'FILLER',
    LIPOLYSIS = 'LIPOLYSIS',
    SKIN_BOOSTER = 'SKIN_BOOSTER',
    COSMETIC = 'COSMETIC',
    HAIR_CARE = 'HAIR_CARE',
    TOXINS = 'TOXINS',
}